feat: рисуем путь на клик
