+++ src/entities/graph/ui/graph.tsx
@@ -1,11 +1,13 @@
 import { Circle, Group, Line } from "react-konva"
 import React, { useMemo } from "react"
 import { TGraph } from "../model/interface"
+import { useGraphContext } from "../../../shared/providers/graph-context/lib/use-graph-context"
 
 export const Graph: React.FC<{ graph: TGraph }> = ({ graph }) => {
+  const { coloredGraph } = useGraphContext()
   const color = useMemo(
-    () => (graph.isFilled ? "red" : undefined),
-    [graph.isFilled],
+    () => (coloredGraph.includes(graph.id) ? "red" : undefined),
+    [coloredGraph, graph.id],
   )
 
   return (
